requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.65.0
wandb>=0.15.0
opencv-python>=4.7.0
matplotlib>=3.7.0
datasets>=2.12.0
huggingface-hub>=0.15.0
argparse>=1.4.0 